Centralizing linear data generator and mllib regression tests to use it.
